requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
faker
django-cors-headers
